Deployment/requirements.txt
gradio
torch
numpy
xgboost
tensorflow
Pillow
opencv-python
torchvision
scikit-image
transformers